# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-percentageof",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `percent% of whole` by composing two float primitives.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. ask `srvcs-floatdivide` for `frac = percent / 100`;\n2. ask `srvcs-floatmultiply` for `result = frac * whole`.\n\nThe constant `100` is a trivial local literal; the headline operations both\ngo through the dependency services. If a dependency is unreachable it reports\nitself degraded (`503`); if a dependency rejects the input it forwards the\n`422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PercentageOfResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "percent",
          "whole"
        ],
        "properties": {
          "percent": {
            "type": "number",
            "format": "double"
          },
          "whole": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "PercentageOfResponse": {
        "type": "object",
        "required": [
          "percent",
          "whole",
          "result"
        ],
        "properties": {
          "percent": {
            "type": "number",
            "format": "double"
          },
          "result": {
            "type": "number",
            "format": "double",
            "description": "`percent% of whole`, as a floating-point number that may be fractional."
          },
          "whole": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}